backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.127.1",
    "httpx",
    "joblib",
    "loguru",
    "mlflow",
    "numpy",
    "pandas",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest",
    "pytest-asyncio",
    "python-dotenv",
    "python-jose>=3.5.0",
    "resend>=2.19.0",
    "scikit-learn",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "argon2-cffi>=25.1.0",
    "google-auth>=2.45.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.3.0",
    "google-api-python-client>=2.187.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
